@@ -2,10 +2,9 @@
 
 namespace ClientAccounts.Models
 {
-    /// <summary>
-    /// Класс для счета (ID счета, ID владельца, тип счета, срок (в месяцах), ставка %, сумма вклада).
-    /// Реализована перегрузка операторов == и != для сравнения счетов с override Equals и GetHashCode-методов.
-    /// </summary>
+    // Класс для счета (ID счета, ID владельца, тип счета, срок (в месяцах), ставка %, сумма вклада).
+    // Реализована перегрузка операторов == и != для сравнения счетов с override Equals и GetHashCode-методов.
+
     class Account
 	{
 		public Guid OwnerID { get; set; } 
