@@ -1,8 +1,7 @@
 ﻿namespace ClientAccounts.Models
 {
-	/// <summary>
-	/// Класс менеджера 
-	/// </summary>
+	// Класс менеджера
+
 	internal class Manager : IUserType
 	{
 		public override string ToString() => "Менеджер";		
