@@ -1,9 +1,8 @@
 ﻿namespace ClientAccounts.ViewModels
 {
-    /// <summary> 
-    /// Класс AccountsReplenishingVM - DataContext для диалог. окна AccountsReplenishingWindow
-    /// Собираем данные о сумме, кот. нужно положить на счет
-    /// </summary>
+    // Класс AccountsReplenishingVM - DataContext для диалог. окна AccountsReplenishingWindow
+    // Собираем данные о сумме, кот. нужно положить на счет
+
     public class AccountsReplenishingVM
 	{
 		public double AddingSum { get; set; }
