@@ -1,9 +1,7 @@
 ﻿namespace ClientAccounts.ViewModels
 {
-    /// <summary> 
-	/// Класс AccountsWithdrawingVM - DataContext для диалог. окна WithdrawingWindow
-	/// Собираем данные о сумме, кот. нужно снять со счета
-	/// </summary>
+	// Класс AccountsWithdrawingVM - DataContext для диалог. окна WithdrawingWindow
+	// Собираем данные о сумме, кот. нужно снять со счета
     public class AccountsWithdrawingVM
 	{
 		public double SubstructingSum {get; set;}		
