@@ -3,12 +3,11 @@ using ClientAccounts.Views;
 
 namespace ClientAccounts.ViewModels
 {
-    /// <summary>
-    /// Класс UserSelectionVM - это DataContext для окна выбора пользователя UserSelectionWindow (менеджер, консультант). 
-	/// В конструктор передается объект ClientsInfoVM.
-	/// При выделении SelectedUser в Combobox-элементе открывается ClientsWindow 
-	/// (при этом обновляем лист клиентов и устанавливаем Changer, какой пользователь делает изменения).
-    /// </summary>
+    // Класс UserSelectionVM - это DataContext для окна выбора пользователя UserSelectionWindow (менеджер, консультант). 
+	// В конструктор передается объект ClientsInfoVM.
+	// При выделении SelectedUser в Combobox-элементе открывается ClientsWindow 
+	// (при этом обновляем лист клиентов и устанавливаем Changer, какой пользователь делает изменения).
+
     class UserSelectionVM
 	{
 		public IUserType[] UserTypes { get; } = new IUserType[] { new Consultant(), new Manager() };
@@ -19,7 +18,7 @@ namespace ClientAccounts.ViewModels
             set
             {
                 ClientsInfoVM.Changer = value;
-                ClientsInfoVM.UpdateClientsList();
+                ClientsInfoVM.LoadClientsList();
                 new ClientsWindow() { DataContext = ClientsInfoVM }.ShowDialog();
             }
         }
